+++ Client/FileSize.h
@@ -13,23 +13,23 @@ enum class SizeUnit
 class FileSize
 {
 private:
-	size_t size;
+	uint64_t size;
 public:
-	FileSize(size_t size = 0, SizeUnit unit = SizeUnit::B);
+	FileSize(uint64_t size = 0, SizeUnit unit = SizeUnit::B);
 
-	size_t to_int() const;
+	uint64_t to_int() const;
 	std::string toString() const;
 
 	FileSize operator+(const FileSize& fileSize);
-	FileSize operator+(const size_t& fileSize);
+	FileSize operator+(const uint64_t& fileSize);
 	FileSize operator+=(const FileSize& fileSize);
-	FileSize operator+=(const size_t& fileSize);
+	FileSize operator+=(const uint64_t& fileSize);
 	FileSize operator-(const FileSize& fileSize);
-	FileSize operator-(const size_t& fileSize);
+	FileSize operator-(const uint64_t& fileSize);
 	FileSize operator-=(const FileSize& fileSize);
-	FileSize operator-=(const size_t& fileSize);
+	FileSize operator-=(const uint64_t& fileSize);
 
-	FileSize operator=(const size_t& fileSize);
+	FileSize operator=(const uint64_t& fileSize);
 	FileSize operator=(const FileSize& fileSize);
 };
 #endif // !FILE_SIZE_H
